Use Tabs2, add loading support
