+++ src/frontend/index.html
@@ -15,9 +15,6 @@
     .pt-overlay .pt-dialog{
       margin-top:2em;
     }
-    .pt-tabs{
-      width:100%;
-    }
     .react-monaco-editor-container{
         border-right:1px solid #EEE;
     }
